requirements.txt
# Core dependencies for SAM2
torch>=2.5.1
torchvision>=0.20.1
numpy>=1.24.4
tqdm>=4.66.1
hydra-core>=1.3.2
iopath>=0.1.10
pillow>=9.4.0

# For notebook usage
matplotlib>=3.9.1
jupyter>=1.0.0
opencv-python>=4.7.0

# For video processing
av>=13.0.0
imageio-ffmpeg

# Optional dependencies for enhanced functionality
# eva-decord>=0.6.1  # May not be available on all systems
requests>=2.25.0
scipy>=1.7.0 